relocate view tests
